// lib/types.ts
export type BlogPost = {
  id: string; // Firestore document ID
  slug: string;
  title: string;
  date: string; // Ou Firestore Timestamp converti en string/Date
  excerpt: string;
  content: string; // HTML content
  imageUrl: string;
  author: string;
  dataAiHint?: string;
  isPublished?: boolean;
  tags?: string[];
  createdAt?: Date;
  updatedAt?: Date;
};